ui update from previous pr
